Refactor tracking functions
