+++ src/module.ts
@@ -2,9 +2,9 @@ import { DataSourcePlugin, DashboardLoadedEvent } from '@grafana/data';
 import { Datasource } from './data/CHDatasource';
 import { ConfigEditor } from './views/CHConfigEditor';
 import { CHQueryEditor } from './views/CHQueryEditor';
-import { CHQuery, CHConfig, BuilderMode } from './types';
+import { CHQuery, CHConfig } from './types';
 import { getAppEvents } from '@grafana/runtime';
-import { trackClickhouseMonitorDashboardLoaded } from 'tracking';
+import { analyzeQueries, trackClickhouseMonitorDashboardLoaded } from 'tracking';
 import pluginJson from './plugin.json';
 
 export const plugin = new DataSourcePlugin<Datasource, CHQuery, CHConfig>(Datasource)
@@ -21,38 +21,12 @@ getAppEvents().subscribe<DashboardLoadedEvent<CHQuery>>(
       return;
     }
 
-    const counters = {
-      sql_queries: 0,
-      builder_queries: 0,
-      builder_table_queries: 0,
-      builder_aggregate_queries: 0,
-      builder_time_series_queries: 0
-    };
-
-    clickhouseQueries.forEach((query) => {
-      switch (query.queryType) {
-        case 'sql':
-          counters.sql_queries++;
-          break;
-        case 'builder':
-          counters.builder_queries++;
-          if (query.builderOptions.mode === BuilderMode.Aggregate) {
-            counters.builder_aggregate_queries++;
-          } else if (query.builderOptions.mode === BuilderMode.List) {
-            counters.builder_table_queries++;
-          } else if (query.builderOptions.mode === BuilderMode.Trend) {
-            counters.builder_time_series_queries++;
-          }
-          break;
-      }
-    });
-
     trackClickhouseMonitorDashboardLoaded({
       clickhouse_plugin_version: pluginJson.info.version,
       grafana_version: grafanaVersion,
       dashboard_id: dashboardId,
       org_id: orgId,
-      ...counters,
+      ...analyzeQueries(clickhouseQueries),
     });
   }
 );
